requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.0.1
PyPDF2==3.0.1
reportlab==4.0.4
ipwhois==1.2.0
lxml==4.9.3
Flask==2.3.3
Werkzeug==2.3.7
reportlab==4.0.4
PyYAML==6.0.1